modules/sc-mesh-secure-deployment/src/nats/cbma/requirements.txt
cffi==1.14.6
cryptography==3.3.2
pycparser==2.20
pyOpenSSL==19.1.0
iniconfig==2.0.0
packaging==23.2
pluggy==1.4.0
pytest==8.0.0
